Added way to retrieve synced_objects_header
